Unify GratingOptics across instruments.
