+++ bundle/edu.gemini.itc/src/main/java/edu/gemini/itc/gmos/Gmos.java
@@ -52,7 +52,7 @@ public abstract class Gmos extends Instrument {
 
     // Keep a reference to the color filter to ask for effective wavelength
     protected Filter _Filter;
-    protected GratingOptics _gratingOptics;
+    protected GmosGratingOptics _gratingOptics;
     protected Detector _detector;
     protected double _sampling;
     protected String _filterUsed;
