+++ bundle/edu.gemini.itc/src/main/java/edu/gemini/itc/gmos/GmosNorth.java
@@ -201,9 +201,7 @@ public class GmosNorth extends Gmos {
 
         if (!(_grating.equals("none"))) {
 
-            _gratingOptics = new GratingOptics(getDirectory() + "/", getPrefix(), _grating, _detector,
-                    // _focalPlaneMaskOffset,
-                    _stringSlitWidth,
+            _gratingOptics = new GmosGratingOptics(getDirectory() + "/" + getPrefix(), _grating, _detector,
                     _centralWavelength,
                     _detector.getDetectorPixels(),//_spectralBinning,
                     _spectralBinning);
