+++ bundle/edu.gemini.itc/src/main/java/edu/gemini/itc/nifs/Nifs.java
@@ -35,7 +35,7 @@ public abstract class Nifs extends Instrument {
 
     // Keep a reference to the color filter to ask for effective wavelength
     protected Filter _Filter;
-    protected GratingOptics _gratingOptics;
+    protected NifsGratingOptics _gratingOptics;
     protected Detector _detector;
     protected double _sampling;
     protected String _filterUsed;
