+++ bundle/edu.gemini.itc/src/main/java/edu/gemini/itc/nifs/NifsNorth.java
@@ -167,9 +167,7 @@ public class NifsNorth extends Nifs {
 
         if (!(_grating.equals("none"))) {
 
-            _gratingOptics = new GratingOptics(getDirectory() + "/", getPrefix(), _grating,
-                    // _focalPlaneMaskOffset,
-                    _stringSlitWidth,
+            _gratingOptics = new NifsGratingOptics(getDirectory() + "/" + getPrefix(), _grating,
                     _centralWavelength,
                     _detector.getDetectorPixels(),//_spectralBinning,
                     1);
